+++ R/targets_misc.R
@@ -33,24 +33,23 @@ list_fastq_files <-
            pattern_R1 = "_R1_",
            pattern_R2 = "_R2_",
            nb_files = Inf) {
-    
-    if(length(list.files(path))==0) {
+    if (length(list.files(path)) == 0) {
       stop("There is no files in the folder ", path)
     }
     list_files <- list.files(path, pattern = pattern, full.names = TRUE)
-    if(length(list_files)==0) {
+    if (length(list_files) == 0) {
       stop("None file in the folder ", path, " match the pattern ", pattern)
     }
     if (paired_end) {
       fnfs <- sort(list_files[grepl(list_files, pattern = pattern_R1)])
-    if(length(fnfs)==0) {
-      stop("None file in the folder ", path, " match the pattern_R1 ", pattern_R1)
-    }
+      if (length(fnfs) == 0) {
+        stop("None file in the folder ", path, " match the pattern_R1 ", pattern_R1)
+      }
       fnrs <-
         sort(list_files[grepl(list_files, pattern = pattern_R2)])
-    if(length(fnrs)==0) {
-      stop("None file in the folder ", path, " match the pattern_R2 ", pattern_R2)
-    }
+      if (length(fnrs) == 0) {
+        stop("None file in the folder ", path, " match the pattern_R2 ", pattern_R2)
+      }
       if (is.finite(nb_files)) {
         fnfs <- fnfs[seq(1, nb_files)]
         fnrs <- fnrs[seq(1, nb_files)]
@@ -409,7 +408,7 @@ sam_data_matching_names <- function(path_sam_data,
     )
     if (verbose) {
       warning(
-        paste(tib_j$raw_sam[is.na(tib_j$raw_fastq)], collapse=" "),
+        paste(tib_j$raw_sam[is.na(tib_j$raw_fastq)], collapse = " "),
         " not matching names from sam_data file."
       )
     }
@@ -422,7 +421,7 @@ sam_data_matching_names <- function(path_sam_data,
     )
     if (verbose) {
       warning(
-        paste(tib_j$raw_fastq[is.na(tib_j$raw_sam)], collapse=" "),
+        paste(tib_j$raw_fastq[is.na(tib_j$raw_sam)], collapse = " "),
         " not matching names from fastq files."
       )
     }